+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Released on Thursday, September 10 2026
 - Updated `DoFocus` and `DoBlur` to apply to form elements
 - Fixed HTML hyperlink pseudo-classes omitting empty `href` and matching `link` elements (#1337)
 - Fixed HTML-namespace element creation losing local-name case (#1327)
+- Fixed changing focus of elements
 
 # 1.8.0